Shannon's suggestions and changes.
