+++ Tests/MotorControlTests/TwoHandNavigationTests.swift
@@ -204,88 +204,6 @@ final class MotorControlNavigationTests: XCTestCase {
     }
 }
 
-final class MotorControlViewModelTests: XCTestCase {
-    
-    func testInstructionStateHand() throws {
-        let left = HandSelection.left.rawValue.uppercased()
-        let right = HandSelection.right.rawValue.uppercased()
-        
-        let instructionStateLeft = MotorControlInstructionState(handInstructionExample, parentId: nil, whichHand: HandSelection.left)
-        guard let title = instructionStateLeft.title, let detail = instructionStateLeft.detail else { throw TestError.nilValue }
-        XCTAssert(title.contains(left))
-        XCTAssert(detail.contains(left))
-        XCTAssertFalse(title.contains(right))
-        XCTAssertFalse(detail.contains(right))
-        
-        let instructionStateRight = MotorControlInstructionState(handInstructionExample, parentId: nil, whichHand: HandSelection.right)
-        guard let title = instructionStateRight.title, let detail = instructionStateRight.detail else { throw TestError.nilValue }
-        XCTAssert(title.contains(right))
-        XCTAssert(detail.contains(right))
-        XCTAssertFalse(title.contains(left))
-        XCTAssertFalse(detail.contains(left))
-        
-        XCTAssertFalse(instructionStateLeft.flippedImage)
-        XCTAssert(instructionStateRight.flippedImage)
-    }
-    
-    func testTappingState() throws {
-        let assessmentState: AssessmentState = .init(AssessmentObject())
-        let tappingStepViewModel: TappingStepViewModel = .init(tappingExample, assessmentState: assessmentState, branchState: assessmentState)
-        var taps = 50
-        for ii in 0..<taps {
-            let tappedButton = getTappingButtonIdentifier(ii)
-            let location = tappedButton == TappingButtonIdentifier.left ? CGPoint(x: 100, y: 500) : CGPoint(x: 200, y: 500)
-            tappingStepViewModel.tappedScreen(uptime: TimeInterval(ii),
-                                              timestamp: TimeInterval(ii),
-                                              currentButton: tappedButton,
-                                              location: location,
-                                              duration: 0.05)
-        }
-        
-        var goodTaps = taps - Int(taps / 3)
-        XCTAssertEqual(tappingStepViewModel.tapCount, goodTaps)
-        XCTAssertEqual(tappingStepViewModel.samples.count, taps)
-
-        tappingStepViewModel.tappedScreen(uptime: TimeInterval(1),
-                                          timestamp: TimeInterval(1),
-                                          currentButton: getTappingButtonIdentifier(taps),
-                                          location: CGPoint(x: 100, y: 500),
-                                          duration: 0.05)
-        taps += 1
-        goodTaps = taps - Int(taps / 3)
-        XCTAssertEqual(tappingStepViewModel.tapCount, goodTaps)
-        XCTAssertEqual(tappingStepViewModel.samples.count, taps)
-        
-        for ii in 0..<taps {
-            tappingStepViewModel.tappedScreen(uptime: TimeInterval(ii),
-                                              timestamp: TimeInterval(ii),
-                                              currentButton: TappingButtonIdentifier.none,
-                                              location: CGPoint(x: 100, y: 500),
-                                              duration: 0.05)
-        }
-        XCTAssertEqual(tappingStepViewModel.tapCount, goodTaps)
-        XCTAssertEqual(tappingStepViewModel.samples.count, taps * 2)
-    }
-}
-
-func getTappingButtonIdentifier(_ ii: Int) -> TappingButtonIdentifier {
-    switch ii % 3 {
-    case 0:
-        return TappingButtonIdentifier.left
-    case 1:
-        return TappingButtonIdentifier.right
-    default:
-        return TappingButtonIdentifier.none
-    }
-}
-
-fileprivate let handInstructionExample = InstructionStepObject(
-    identifier: "instruction",
-    title: "Tap with your %@ hand",
-    detail: "Alternate tapping the buttons that appear with your index and middle fingers on your %@ HAND. Keep tapping for 30 seconds as fast as you can.")
-
-fileprivate let tappingExample = TappingNodeObject(identifier: "tappingExample")
-
 class TestNavigationState : NavigationState {
     let assessmentState: AssessmentState
